split to proximity.js for unit tests
