+++ lib/geocode.js
@@ -1,7 +1,6 @@
 var sm = new (require('sphericalmercator'))(),
     ops = require('./util/ops'),
     phrasematch = require('./phrasematch'),
-    SphericalMercator = require('sphericalmercator'),
     context = require('./context'),
     termops = require('./util/termops'),
     spatialmatch = require('./spatialmatch'),
@@ -9,10 +8,9 @@ var sm = new (require('sphericalmercator'))(),
     queue = require('queue-async'),
     feature = require('./util/feature'),
     Relev = require('./util/relev'),
-    distance = require('turf-distance'),
-    Point = require('turf-point');
+    proximity = require('./util/proximity');
 var dedupe = require('./util/dedupe');
-var sm = new SphericalMercator();
+
 
 module.exports = function(geocoder, query, options, callback) {
     options = options || {};
@@ -213,8 +211,6 @@ function forwardGeocode(geocoder, query, options, callback) {
     function spatialmatchComplete(err, matched) {
         if (err) return callback(err);
 
-        function toCenter(bbox) { return sm.forward([bbox[0] - bbox[3], bbox[1] - bbox[3]]); }
-
         // Debug:
         // Determine what position the traced feature is at in the spatialmatch
         // resultset and decode the relev object. Provide the stack that calculated
@@ -243,28 +239,7 @@ function forwardGeocode(geocoder, query, options, callback) {
         //If proximity is set, terms that tie in relev will be ordered by distance
         //This is performed as verify match has a cutoff of the first 40 results
         if (options.proximity) {
-            //Setup grid lookup - each layer is an object in the array with id->grid
-            var gridlookup = [];
-            for (var layer_it = 0; layer_it < grids.length; layer_it++) {
-                gridlookup[layer_it] = {};
-                for (var grid_it = 0; grid_it < grids[layer_it].length; grid_it++) {
-                    tmpGrid = ops.grid(grids[layer_it][grid_it]);
-                    gridlookup[layer_it][tmpGrid.id] = tmpGrid;
-                }
-            }
-            //The centre of each tile is calculated in merc and then sorted
-            var proxSM = sm.forward(options.proximity);
-            matched.results.sort(function(a, b) {
-                var res = [a,b].map(function(ele) { return new Relev(ele); });
-                if (res[0].relev !== res[1].relev) return 0; //Only sort with same relev
-                res = res.map(function(ele) {
-                    var center = toCenter(sm.bbox(gridlookup[ele.idx][ele.id].x, gridlookup[ele.idx][ele.id].y, zooms[ele.idx]));
-                    return Math.sqrt(Math.pow(center[0] - proxSM[0],2) + Math.pow(center[1] - proxSM[1],2));
-                });
-                if (res[0] < res[1]) return -1; //a is closer
-                else if (res[1] < res[0]) return 1; //b is closer
-                else return 0; //same
-            });
+            matched = proximity.coarse(matched, grids, zooms, options);
         }
 
         verifymatch(queryData.query, stats, geocoder, matched, options, function(err, contexts) {
@@ -305,18 +280,7 @@ function forwardGeocode(geocoder, query, options, callback) {
 
             //Second fine-grained proximity sort
             if (options.proximity) {
-                queryData.features.sort(function(a, b) {
-                    //Only sort features with same relev/idx level
-                    if (a.relevance !== b.relevance) return 0;
-                    if (a.id.split('.')[0] !== b.id.split('.')[0]) return 0;
-
-                    var res = [a, b].map(function(ele) {
-                        return distance(Point(options.proximity), Point(ele.center), 'miles');
-                    });
-                    if (res[0] < res[1]) return -1; //a is closer
-                    else if (res[1] < res[0]) return 1; //b is closer
-                    else return 0; //same
-                });
+                queryData = proximity.fine(queryData, options);
             }
 
             if (options.debug) queryData.debug = options.debug;
